requirements.txt
fastapi
uvicorn
oracledb
sqlalchemy
pydantic
python-dotenv